refactor(cypress): rename config files and imports to use cjs so cypress runs locally
