+++ cypress.config.cjs
@@ -2,10 +2,10 @@ const { defineConfig } = require('cypress');
 const fs = require('fs');
 const path = require('path');
 
-const benchmarkPlugin = require('./e2e/cypress/plugins/benchmark/index');
-const readProvisions = require('./e2e/cypress/plugins/readProvisions');
-const smtpTester = require('./e2e/cypress/plugins/smtpTester');
-const typescriptPreprocessor = require('./e2e/cypress/plugins/typescriptPreprocessor');
+const benchmarkPlugin = require('./e2e/cypress/plugins/benchmark/index.cjs');
+const readProvisions = require('./e2e/cypress/plugins/readProvisions.cjs');
+const smtpTester = require('./e2e/cypress/plugins/smtpTester.cjs');
+const typescriptPreprocessor = require('./e2e/cypress/plugins/typescriptPreprocessor.cjs');
 
 module.exports = defineConfig({
   projectId: 'zb7k1c',
