+++ e2e/cypress/plugins/benchmark/CDPDataCollector.cjs
