+++ e2e/cypress/plugins/benchmark/formatting.cjs
