+++ e2e/cypress/plugins/extendConfig.cjs
