+++ e2e/cypress/support/e2e.js
@@ -1,4 +1,4 @@
-require('./commands');
+import './commands.js';
 
 Cypress.Screenshot.defaults({
   screenshotOnRunFailure: false,
